fix. 도전 퇴장시, 시작일이 지나고 나서 퇴장하면 totalpoint만큼 차감되도록 설정.

+++ src/challenges/services/challenges.service.ts
@@ -193,6 +193,19 @@ export class ChallengesService {
     if (!challenge) {
       throw new NotFoundException('해당 도전 게시글이 조회되지 않습니다.');
     }
+    const date = challenge.startDate;
+    const startdate = new Date(date);
+    const today = new Date();
+    const user = await this.userRepository.getUserById(userId);
+
+    //
+    if (startdate <= today) {
+      user.point - challenge.totalPoint;
+      return {
+        message: '도전 시작일 전 퇴장을 하셨으므로, total포인트가 차감됩니다.',
+      };
+    }
+
     await this.challengersRepository.deleteChallenger(challenge.id, userId);
   }
 
@@ -240,7 +253,7 @@ export class ChallengesService {
     // await this.challengesRepository.inviteChallenge(challengeId, friend);
   }
 
-  // 도전 초대수락
+  // 도전 초대수락 (상우)
   async acceptChallenge(
     challengeId: number,
     body: ResponseChallengeDto,
